adding res.status(200)

+++ utils/instagram.js
@@ -2,27 +2,39 @@ const { downloader: Downloader } = require("instagram-url-downloader");
 const fs = require("fs");
 const path = require("path");
 const { dirRoot } = require("../config");
+const { escapeTags } = require("./format-string");
 
 const checkDirExist = (pathname) => {
     let dir = path.dirname(pathname);
     if(dir){
         if(!fs.existsSync(dir)){
             fs.mkdirSync(dir, { recursive: true });
+            console.log(`Make directory: ${dir}`);
         }
     }
 }
 
-const removeMedias = () => {
-    const medias = path.join(dirRoot, "Medias");
-    if(!fs.existsSync(medias)) return;
-    fs.rmSync(medias, { recursive: true, force: true });
+const fileExist = (file) => {
+    if(!file) return false;
+    return fs.existsSync(file);
+}
+
+const removeForce = (file) => {
+    if(!fs.existsSync(file)){
+        console.log(`removeForce(): File or directory not exists`);
+        return false;
+    }
+    fs.rmSync(file, { recursive: true, force: true });
+    console.log(`Delete: ${file}`);
+    return true;
 }
 
 const getMedia = (url, callback) => {
-    removeMedias();
     const downloader = new Downloader(url);
     const media = downloader.Media;
+    media.caption.text = escapeTags(media.caption.text).replace(/\n/gim, "<br>");
     callback(media);
+    console.log(media);
 }
 
 const downloadMedia = (media) => {
@@ -34,11 +46,18 @@ const downloadMedia = (media) => {
 }
 
 const moveDownloaded = (file, media) => {
-    const newFilename = media.ID + path.extname(file);
-    const newpath = path.join(dirRoot, "views/downloads") + "/" + newFilename;
-    checkDirExist(newpath);
-    fs.renameSync(file, newpath);
-    return newpath;
+    return new Promise((resolve, reject) => {
+        if(!fileExist(file)){
+            reject("Media not exists");
+            return false;
+        }
+        const newFilename = media.ID + path.extname(file);
+        const newpath = path.join(dirRoot, "views", "downloads", newFilename);
+        checkDirExist(newpath);
+        fs.renameSync(file, newpath);
+        console.log(`Move from: ${file}\nMove to: ${newpath}`);
+        resolve(newpath)
+    });
 }
 
-module.exports = { getMedia, downloadMedia, moveDownloaded };
+module.exports = { getMedia, downloadMedia, moveDownloaded, removeForce };
